requirements.txt
Flask==3.0.3
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.153.0
python-dotenv==1.0.1
gunicorn==23.0.0
